% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_to_iso.R
\name{measure_to_iso}
\alias{measure_to_iso}
\title{Convert a discrete probability distribution to an isotonic function}
\usage{
measure_to_iso(mu, n = NULL, tol = .Machine$double.eps * 10000)
}
\arguments{
\item{mu}{A named list containing two vectors of the same length, \code{`vals`} and \code{`probs`}.
mu represents a discrete probability distribution.}

\item{n}{integer number of equidistant points on [min(mu$vals), max(mu$vals)] which will be the support of the
returned isotonic function}

\item{tol}{tolerance for shifting cumulative probabilities}
}
\value{
Vector representing isotonic function values
}
\description{
Convert a discrete probability distribution to an isotonic function
}
