% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_to_iso.R
\name{measure_to_smooth_iso}
\alias{measure_to_smooth_iso}
\title{Convert a discrete probability distribution to a smooth mostly-isotonic function}
\usage{
measure_to_smooth_iso(
  mu,
  n,
  kernel = "normal",
  bandwidth = 0.1,
  tol = .Machine$double.eps * 10
)
}
\arguments{
\item{mu}{A named list containing two vectors of the same length, \code{`vals`} and \code{`probs`}.
mu represents a discrete probability distribution.}

\item{n}{integer number of equidistant points on [min(mu$vals), max(mu$vals)] which will be the support of the
returned isotonic function}

\item{kernel}{the kernel to be used. Can be abbreviated. Type \code{?ksmooth} for details.}

\item{bandwidth}{the bandwidth. The kernels are scaled so that their
quartiles (viewed as probability densities) are at +/- 0.25*bandwidth. Type \code{?ksmooth} for details.}

\item{tol}{tolerance for shifting cumulative probabilities}
}
\value{
Vector representing isotonic function values
}
\description{
Convert a discrete probability distribution to a smooth mostly-isotonic function
}
