% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbernoulli_custom}
\alias{pbernoulli_custom}
\title{Discrete distribution with 2 values.}
\usage{
pbernoulli_custom(x, a, b, p)
}
\arguments{
\item{x}{observations.}

\item{a}{first value.}

\item{b}{second value.}

\item{p}{Let X ~ pbernoulli_custom. Then p = Probability(X=a).}
}
\description{
Discrete distribution with 2 values.
}
