% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbernoulli_custom}
\alias{rbernoulli_custom}
\title{Discrete distribution with 2 values.}
\usage{
rbernoulli_custom(n, a = -1, b = 1, p = 0.5)
}
\arguments{
\item{n}{number of observations.}

\item{a}{first value.}

\item{b}{second value.}

\item{p}{Let X ~ pbernoulli_custom. Then p = Probability(X=a).}
}
\description{
Discrete distribution with 2 values.
}
