% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UncoupledIsoReg.R
\docType{package}
\name{UncoupledIsoReg-package}
\alias{UncoupledIsoReg}
\alias{UncoupledIsoReg-package}
\title{UncoupledIsoReg: Package for Uncoupled Isotonic Regression}
\description{
Based on algorithm from arXiv:1806.10648 by Rigollet and Weed.
}
\details{
\code{UncoupledIsoReg} is a package for uncoupled isotonic regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jan-imbi.github.io/UncoupledIsoReg}
}

}
\author{
\strong{Maintainer}: Jan Meis \email{meis@imbi.uni-heidelberg.de}

}
\keyword{internal}
