% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Wp}
\alias{Wp}
\title{Minimize the entropic Wasserstein distance}
\usage{
Wp(mu_vals, mu_probs, nu_vals, nu_probs, p)
}
\arguments{
\item{mu_vals, nu_vals}{numeric vector of values of discrete distribution}

\item{mu_probs, nu_probs}{numeric vector of probabilities of discrete distribution}

\item{p}{exponent of L_p distance}
}
\description{
Minimize the entropic Wasserstein distance
}
\examples{
Wp(c(1,2,3), c(1/3, 1/3, 1/3), c(2,3,4), c(1/4, 1/4, 1/2), 1)
}
